+++ python/cudf_polars/cudf_polars/dsl/expressions/rolling.py
@@ -1,0 +1,38 @@
+# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES.
+# SPDX-License-Identifier: Apache-2.0
+# TODO: remove need for this
+# ruff: noqa: D101
+"""Rolling DSL nodes."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+from cudf_polars.dsl.expressions.base import Expr
+
+if TYPE_CHECKING:
+    import pylibcudf as plc
+
+
+class RollingWindow(Expr):
+    __slots__ = ("options", "children")
+    _non_child = ("dtype", "options")
+    children: tuple[Expr]
+
+    def __init__(self, dtype: plc.DataType, options: Any, agg: Expr) -> None:
+        super().__init__(dtype)
+        self.options = options
+        self.children = (agg,)
+        raise NotImplementedError("Rolling window not implemented")
+
+
+class GroupedRollingWindow(Expr):
+    __slots__ = ("options", "children")
+    _non_child = ("dtype", "options")
+    children: tuple[Expr, ...]
+
+    def __init__(self, dtype: plc.DataType, options: Any, agg: Expr, *by: Expr) -> None:
+        super().__init__(dtype)
+        self.options = options
+        self.children = (agg, *by)
+        raise NotImplementedError("Grouped rolling window not implemented")
